Rename publishers module to qos
